Update to profile, application, and new logout page
Application update: will now check to see if a session is in process and will fill a global variable "g" for use on other html pages

Profile update: tested global variable "g" and how it outputs mongoDB object attributes

Logout html: going to this link will auto logout the current session. Button is needed to be made that will replace Login button when session is in process.